pyproject.toml
[build-system]
requires = ["setuptools>=80.9.0", "setuptools-scm[simple]>=9.2.0"]
build-backend = "setuptools.build_meta"

[project]
name = "rtl_buddy"
dynamic = ["version"]
requires-python = ">=3.11"
dependencies = [
  "ruamel.yaml >= 0.18.15",
  "typer >= 0.19.2",
  "click >= 8.3.0",
  "pyserde[yaml] >= 0.27.0",
  "rich >= 14.0.0"
]

authors = [
  { name = "Zhi-Hern Loh"},
  { name = "Xie Jiacheng"}
]

description = "RTL Buddy is a build-system with testplan, regression, workflow support for Verilog RTL codebases."
readme = "README.md"
license = "BSD-3-Clause"
license-files = ["LICENSE"]
keywords = [ "SystemVerilog", "Verilog", "ASIC", "FPGA" ]

classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Scientific/Engineering :: Electronic Design Automation (EDA)",
  "Operating System :: POSIX :: Linux",
  "Operating System :: MacOS",
]

[project.urls]
Source = "https://github.com/rtl-buddy/rtl_buddy"
Documentation = "https://rtl-buddy.github.io/rtl_buddy/"
Tracker = "https://github.com/rtl-buddy/rtl_buddy/issues"

[tool.setuptools_scm]

[tool.setuptools.package-data]
"rtl_buddy.skill" = ["SKILL.md", "gitignore_snippet.txt"]

[dependency-groups]
docs = [
  "mkdocs>=1.6.0",
  "mkdocs-material>=9.5.0",
  "mike>=2.1.0",
]
lint = [
  "ruff>=0.11.0",
]
test = [
  "pytest>=8.4.0",
]
dev = [
  { include-group = "docs" },
  { include-group = "lint" },
  { include-group = "test" },
]

[project.scripts]
rtl_buddy = "rtl_buddy.__main__:main"
rtl-buddy = "rtl_buddy.__main__:main"
rb = "rtl_buddy.__main__:main"

[tool.ruff]
line-length = 88
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["F", "E7", "E9"]
